LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-min-width-composition-expected.txt


FAIL Compositing: property <min-width> underlying [100px] from add [max-content] to add [100px] at (-0.3) should be [calc-size(max-content, 70px + size * 1.3)] assert_equals: expected "0px " but got "max - content "
FAIL Compositing: property <min-width> underlying [100px] from add [max-content] to add [100px] at (0) should be [calc-size(max-content, 100px + size * 1)] assert_equals: expected "0px " but got "max - content "
FAIL Compositing: property <min-width> underlying [100px] from add [max-content] to add [100px] at (0.5) should be [calc-size(max-content, 150px + size * 0.5)] assert_equals: expected "0px " but got "200px "
FAIL Compositing: property <min-width> underlying [100px] from add [max-content] to add [100px] at (1) should be [calc-size(max-content, 200px + size * 0)] assert_equals: expected "0px " but got "200px "
FAIL Compositing: property <min-width> underlying [100px] from add [max-content] to add [100px] at (1.5) should be [calc-size(max-content, 250px + size * -0.5)] assert_equals: expected "0px " but got "200px "
PASS Compositing: property <min-width> underlying [min-content] from add [fit-content] to add [max-content] at (-0.3) should be [fit-content]
PASS Compositing: property <min-width> underlying [min-content] from add [fit-content] to add [max-content] at (0) should be [fit-content]
PASS Compositing: property <min-width> underlying [min-content] from add [fit-content] to add [max-content] at (0.5) should be [max-content]
PASS Compositing: property <min-width> underlying [min-content] from add [fit-content] to add [max-content] at (1) should be [max-content]
PASS Compositing: property <min-width> underlying [min-content] from add [fit-content] to add [max-content] at (1.5) should be [max-content]
FAIL Compositing: property <min-width> underlying [max-content] from add [100px] to add [min-content] at (-0.3) should be [calc-size(min-content, 130px + size * -0.3)] assert_equals: expected "0px " but got "100px "
FAIL Compositing: property <min-width> underlying [max-content] from add [100px] to add [min-content] at (0) should be [calc-size(min-content, 100px + size * 0)] assert_equals: expected "0px " but got "100px "
FAIL Compositing: property <min-width> underlying [max-content] from add [100px] to add [min-content] at (0.5) should be [calc-size(min-content, 50px + size * 0.5)] assert_equals: expected "0px " but got "min - content "
FAIL Compositing: property <min-width> underlying [max-content] from add [100px] to add [min-content] at (1) should be [calc-size(min-content, 0px + size * 1)] assert_equals: expected "0px " but got "min - content "
FAIL Compositing: property <min-width> underlying [max-content] from add [100px] to add [min-content] at (1.5) should be [calc-size(min-content, -50px + size * 1.5)] assert_equals: expected "0px " but got "min - content "
